# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    ListApplicationsResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/PdApplicationDto"
...
    PdApplicationDto:
      type: object
      properties:
        id:
          type: string
        repoId:
          type: string
        name:
          type: string
        environmentId:
          type: string
        environmentName:
          type: string
        availableOnEnv:
          type: boolean
        branch:
          type: string
        healthPath:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
...
    RetirementResponse:
      type: object
      properties:
        applicationId:
          type: string
        applicationName:
          type: string
        environmentId:
          type: string
        currentDeploymentId:
          type: string
        previousStatus:
          type: string
        decommissionedIds:
          type: array
          items:
            type: string
    ScaleRequest:
      type: object
      properties:
        replicas:
          type: integer
          format: int32
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
  /deployments/api/applications:
    get:
      summary: Every application deployed here — environment applications and platform
        services
      operationId: listApplications
      responses:
        "200":
          description: "The applications, one entry per tier"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListApplicationsResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Application Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
  /deployments/api/applications/{applicationId}/decommission:
    post:
      summary: "Retire an application — its current row becomes DECOMMISSIONED, nothing\
        \ is deleted"
      operationId: decommissionApplication
      parameters:
      - name: applicationId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Retired; the rows it settled are named
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetirementResponse"
        "400":
          description: Not an application id
        "404":
          description: Nothing has ever been deployed for this application
        "409":
          description: "It is still deployed, or the deploy worker still has work\
            \ for it"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Application Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
  /deployments/api/applications/{applicationId}/restart:
    post:
      summary: "Restart an application in place — its tasks are replaced, its deployment\
        \ is not"
      operationId: restartApplication
      parameters:
      - name: applicationId
        in: path
        required: true
        schema:
          type: string
      responses:
        "202":
          description: Queued on the deploy worker
        "400":
          description: Not an application id
        "404":
          description: Nothing has ever been deployed for this application
        "409":
          description: Its newest deployment never reached the orchestrator
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Application Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
  /deployments/api/applications/{applicationId}/scale:
    post:
      summary: "Scale an application's workload — 0 stops it, 1 runs it"
      operationId: scaleApplication
      parameters:
      - name: applicationId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScaleRequest"
        required: true
      responses:
        "202":
          description: Queued on the deploy worker
        "400":
          description: "No replica count, a negative one, or one above the single\
            \ task every application here is deployed as"
        "404":
          description: Nothing has ever been deployed for this application
        "409":
          description: Its newest deployment never reached the orchestrator
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Application Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
...
info:
  title: qits-platform-deployments API
  version: 2026.1008.164327